requirements-deploy.txt
-r requirements.txt
PyYAML~=3.12
mysqlclient~=1.3
uwsgi~=2.0
